Revert code moves since they don't make sense in the new Core structure
